\name{mongo.aggregate.stats}
\alias{mongo.aggregate.stats}
\title{Compute Summary Statistics of Mongo Subsets}
\usage{
  mongo.aggregate.stats(mongo, ns, keys, by)
}
\description{
  Splits the mongo data into subsets, computes summary
  statistics for eachm and returns the result in a
  convenient form.
}
\seealso{
  \code{\link{aggregate}}
}

