% File rmongodb/man/mongo.binary.binary.Rd
\name{mongo.binary.binary}
\alias{mongo.binary.binary}
\title{BSON binary data subtype constant for standard binary data}
\description{
BSON binary data subtype constant for standard binary data.
}
\usage{
mongo.binary.binary
}
\value{
0L
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
