% File rmongodb/man/mongo.binary.md5.Rd
\name{mongo.binary.md5}
\alias{mongo.binary.md5}
\title{BSON binary data subtype constant for md5 data}
\description{
BSON binary data subtype constant for md5 data.
}
\usage{
mongo.binary.md5
}
\value{
5L
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
