% File rmongodb/man/mongo.binary.user.Rd
\name{mongo.binary.user}
\alias{mongo.binary.user}
\title{BSON binary data subtype constant for user data}
\description{
BSON binary data subtype constant for user data.
}
\usage{
mongo.binary.user
}
\value{
128L
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
