% File rmongodb/man/mongo.binary.uuid.Rd
\name{mongo.binary.uuid}
\alias{mongo.binary.uuid}
\title{BSON binary data subtype constant for uuid data}
\description{
BSON binary data subtype constant for uuid data.
}
\usage{
mongo.binary.uuid
}
\value{
4L
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
