% File rmongodb/man/mongo.bson.binary.Rd
\name{mongo.bson.binary}
\alias{mongo.bson.binary}
\title{BSON data type constant for a binary data value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(5L) to indicate that the value pointer to by an iterator is binary data.
}
\usage{
mongo.bson.binary
}
\value{
5L
}
\seealso{
\code{\link{mongo.bson.iterator.type}}
\code{\link{mongo.bson.iterator.next}}
\link{mongo.bson}
}
