% File rmongodb/man/mongo.bson.code.w.scope.Rd
\name{mongo.bson.code.w.scope}
\alias{mongo.bson.code.w.scope}
\title{BSON data type constant for a code with scope value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(15L) to indicate that the value pointer to by an iterator is a javascript with a scope.
}
\usage{
mongo.bson.code.w.scope
}
\value{
15L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}
}
