% File rmongodb/man/mongo.bson.date.Rd
\name{mongo.bson.date}
\alias{mongo.bson.date}
\title{BSON data type constant for a date value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(9L) to indicate that the value pointer to by an iterator is a date/time.
}
\usage{
mongo.bson.date
}
\value{
9L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
