% File rmongodb/man/mongo.bson.double.Rd
\name{mongo.bson.double}
\alias{mongo.bson.double}
\title{BSON data type constant for a double value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(1L) to indicate that the value pointer to by an iterator is a double.
}
\usage{
mongo.bson.double
}
\value{
1L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
