% File rmongodb/man/mongo.bson.eoo.Rd
\name{mongo.bson.eoo}
\alias{mongo.bson.eoo}
\title{BSON data type constant for 'End Of Object'}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(0L) at the end of the object when there are no more fields through which to iterate.
}
\usage{
mongo.bson.eoo
}
\value{
0L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
