% File rmongodb/man/mongo.bson.int.Rd
\name{mongo.bson.int}
\alias{mongo.bson.int}
\title{BSON data type constant for a integer value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(16L) to indicate that the value pointer to by an iterator is a integer (32-bit).
}
\usage{
mongo.bson.int
}
\value{
16L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
