% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.bson.long}
\alias{mongo.bson.long}
\title{BSON data type constant for a long value}
\format{\preformatted{ int 18
}}
\usage{
mongo.bson.long
}
\value{
18L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this constant (18L) to
indicate that the value pointer to by an iterator is a long integer (64
bits).
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr \link{mongo.bson}.
}
\keyword{datasets}

