% File rmongodb/man/mongo.bson.null.Rd
\name{mongo.bson.null}
\alias{mongo.bson.null}
\title{BSON data type constant for a null value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(10L) to indicate that the value pointer to by an iterator is a null.
}
\usage{
mongo.bson.null
}
\value{
10L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
