% File rmongodb/man/mongo.bson.object.Rd
\name{mongo.bson.object}
\alias{mongo.bson.object}
\title{BSON data type constant for a subobject value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(3L) to indicate that the value pointer to by an iterator is a subobject.
}
\usage{
mongo.bson.object
}
\value{
3L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
