% File rmongodb/man/mongo.bson.oid.Rd
\name{mongo.bson.oid}
\alias{mongo.bson.oid}
\title{BSON data type constant for a oid value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(7L) to indicate that the value pointer to by an iterator is a oid (Object ID).
}
\usage{
mongo.bson.oid
}
\value{
7L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
