% File rmongodb/man/mongo.bson.regex.Rd
\name{mongo.bson.regex}
\alias{mongo.bson.regex}
\title{BSON data type constant for a regex value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(11L) to indicate that the value pointer to by an iterator is a regular expression.
}
\usage{
mongo.bson.regex
}
\value{
11L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
