% File rmongodb/man/mongo.bson.string.Rd
\name{mongo.bson.string}
\alias{mongo.bson.string}
\title{BSON data type constant for a string value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(2L) to indicate that the value pointer to by an iterator is a string.
}
\usage{
mongo.bson.string
}
\value{
2L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
