% File rmongodb/man/mongo.bson.undefined.Rd
\name{mongo.bson.undefined}
\alias{mongo.bson.undefined}
\title{BSON data type constant for a undefined value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(6L) to indicate that the value pointer to by an iterator is a undefined.
}
\usage{
mongo.bson.undefined
}
\value{
6L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
