% File rmongodb/man/mongo.bson_arrsy.Rd
\name{mongo.bson_arrsy}
\alias{mongo.bson_arsry}
\title{BSON data type constant for an array}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(4L) to indicate that the value pointer to by an iterator is an array (containing child values).
}
\usage{
mongo.bson_array
}
\value{
3L
}
\seealso{
\code{\link{mongo.bson.iterator.type}}
\code{\link{mongo.bson.iterator.next}}
\link{mongo.bson}
}
