% File rmongodb/man/mongo.bson_double.Rd
\name{mongo.bson_double}
\alias{mongo.bson_double}
\title{BSON data type constant for a double value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(1L) to indicate that the value pointer to by an iterator is a double.
}
\usage{
mongo.bson_double
}
\value{
1L
}
\seealso{
\code{\link{mongo.bson.iterator.type}}
\code{\link{mongo.bson.iterator.next}}
\link{mongo.bson}
}
