% File rmongodb/man/mongo.bson_string.Rd
\name{mongo.bson_string}
\alias{mongo.bson_string}
\title{BSON data type constant for a string value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(2L) to indicate that the value pointer to by an iterator is a string.
}
\usage{
mongo.bson_string
}
\value{
2L
}
\seealso{
\code{\link{mongo.bson.iterator.type}}
\code{\link{mongo.bson.iterator.next}}
\link{mongo.bson}
}
