% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mongo.cursor.to.rlist}
\alias{mongo.cursor.to.rlist}
\title{Convert Mongo Cursor Object to List so that each element of resulting list represents document in source collection.}
\usage{
mongo.cursor.to.rlist(cursor, keep.ordering = TRUE)
}
\arguments{
\item{cursor}{(\link{mongo.cursor}) A mongo.cursor object returned from
\code{\link{mongo.find}()}.}

\item{keep.ordering}{should the records be returned at the same order as fetched from cursor (if sorting was specified in query)?
For speed try to set this parameter to FALSE. This will prevent sorting after fetching from cursor.}
}
\value{
An R \link{list} object.
}
\description{
Converts a mongo cursor object to a list by interating over all cursor objects and combining them.
In contrast to \link{mongo.cursor.to.list} it doesn't make any data coercion!
Just one-to-one mapping with documents in source collection.
}
\details{
Function uses environments to avoid extra copying. It is much faster than \link{mongo.cursor.to.list}
and in next release will replace it.
}
\examples{
mongo <- mongo.create()
if (mongo.is.connected(mongo)) {
    buf <- mongo.bson.buffer.create()
    mongo.bson.buffer.append(buf, "age", 22L)
    query <- mongo.bson.from.buffer(buf)

    # Find the first 100 records
    #    in collection people of database test where age == 22
    cursor <- mongo.find(mongo, "test.people", query, limit=100L)

    res <- mongo.cursor.to.rlist(cursor)

}
}
\seealso{
\code{\link{mongo.find}},\cr \code{\link{mongo.bson.to.list}}.
}

