% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.find.await.data}
\alias{mongo.find.await.data}
\title{mongo.find flag constant - await data}
\format{\preformatted{ int 32
}}
\usage{
mongo.find.await.data
}
\value{
32L
}
\description{
\code{\link{mongo.find}()} flag constant - await data.
}
\keyword{datasets}

