\docType{data}
\name{mongo.find.cursor.tailable}
\alias{mongo.find.cursor.tailable}
\title{mongo.find flag constant - cursor tailable}
\format{int 2}
\usage{
  mongo.find.cursor.tailable
}
\value{
  2L
}
\description{
  \code{\link{mongo.find}()} flag constant - cursor
  tailable.
}
\keyword{datasets}

