% File rmongodb/man/mongo.find.no.cursor.timeout.Rd
\name{mongo.find.no.cursor.timeout}
\alias{mongo.find.no.cursor.timeout}
\title{mongo.find flag constant - no cursor timeout}
\description{
\code{\link{mongo.find}()} flag constant - no cursor timeout.
}
\usage{
mongo.find.no.cursor.timeout
}
\value{
16L
}
