% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.find.oplog.replay}
\alias{mongo.find.oplog.replay}
\title{mongo.find flag constant - oplog replay}
\format{\preformatted{ int 8
}}
\usage{
mongo.find.oplog.replay
}
\value{
8L
}
\description{
\code{\link{mongo.find}()} flag constant - oplog replay.
}
\keyword{datasets}

