\docType{data}
\name{mongo.find.partial.results}
\alias{mongo.find.partial.results}
\title{mongo.find flag constant - partial results}
\format{int 128}
\usage{
  mongo.find.partial.results
}
\value{
  128L
}
\description{
  \code{\link{mongo.find}()} flag constant - partial
  results.
}
\keyword{datasets}

