% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.find.slave.ok}
\alias{mongo.find.slave.ok}
\title{mongo.find flag constant - slave ok}
\format{\preformatted{ int 4
}}
\usage{
mongo.find.slave.ok
}
\value{
4L
}
\description{
\code{\link{mongo.find}()} flag constant - slave ok.
}
\keyword{datasets}

