% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.index.background}
\alias{mongo.index.background}
\title{mongo.index.create flag constant - background}
\format{\preformatted{ int 8
}}
\usage{
mongo.index.background
}
\value{
8L
}
\description{
\code{\link{mongo.index.create}()} flag constant - background.
}
\keyword{datasets}

