% File rmongodb/man/mongo.index.drop.dups.Rd
\name{mongo.index.drop.dups}
\alias{mongo.index.drop.dups}
\title{mongo.index.create flag constant - drop duplicate keys}
\description{
\code{\link{mongo.index.create}()} flag constant - drop duplicate keys.
}
\usage{
mongo.index.drop.dups
}
\value{
4L
}
