% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.index.sparse}
\alias{mongo.index.sparse}
\title{mongo.index.create flag constant - sparse}
\format{\preformatted{ int 16
}}
\usage{
mongo.index.sparse
}
\value{
16L
}
\description{
\code{\link{mongo.index.create}()} flag constant - sparse.
}
\keyword{datasets}

