% File rmongodb/man/mongo.update.basic.Rd
\name{mongo.update.basic}
\alias{mongo.update.basic}
\title{mongo.update() flag constant for performing a basic update}
\description{
Flag to \code{\link{mongo.update}()} (4L): Perform a basic update.
 }
\usage{
mongo.update.basic
}
\value{
4L
}
\seealso{
\code{\link{mongo.update}},\cr
\code{\link{mongo.update.multi}}\cr
\code{\link{mongo.update.upsert}}
}
