% File rmongodb/man/mongo.update.upsert.Rd
\name{mongo.update.upsert}
\alias{mongo.update.upsert}
\title{mongo.update() flag constant for an upsert}
\description{
Flag to \code{\link{mongo.update}()} (1L): insert ObjNew into the database if no record matching criteria is found.
}
\usage{
mongo.update.upsert
}
\value{
1L
}
\seealso{
\code{\link{mongo.update}},\cr
\code{\link{mongo.update.multi}},\cr
\code{\link{mongo.update.basic}}.
}
