\name{rmongo-package}
\alias{rmongo-package}
\alias{rmongo}
\docType{package}
\title{
R-MongoDB driver
}
\description{
Provides an interface to MongoDB for R
}
\details{
\tabular{ll}{
Package: \tab rmongo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-27\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
Overview
}
\author{
Gerald Lindsly

Maintainer: Gerald Lindsly
gerald.lindsly@gmail.com
}
\references{
http://www.mongodb.org
}
\keyword{ package }
\examples{
}
