\docType{data}
\name{mongo.binary.binary}
\alias{mongo.binary.binary}
\title{BSON binary data subtype constant for standard binary data}
\format{\preformatted{ int 0
}}
\usage{
mongo.binary.binary
}
\value{
0L
}
\description{
BSON binary data subtype constant for standard binary data.
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
\keyword{datasets}

