\docType{data}
\name{mongo.binary.function}
\alias{mongo.binary.function}
\title{BSON binary data subtype constant for function data}
\format{int 1}
\usage{
  mongo.binary.function
}
\value{
  1L
}
\description{
  BSON binary data subtype constant for function data.
}
\seealso{
  \code{\link{mongo.bson.buffer.append.raw}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

