% File rmongodb/man/mongo.binary.function.Rd
\name{mongo.binary.function}
\alias{mongo.binary.function}
\title{BSON binary data subtype constant for function data}
\description{
BSON binary data subtype constant for function data.
}
\usage{
mongo.binary.function
}
\value{
1L
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
