\docType{data}
\name{mongo.binary.md5}
\alias{mongo.binary.md5}
\title{BSON binary data subtype constant for md5 data}
\format{int 5}
\usage{
  mongo.binary.md5
}
\value{
  5L
}
\description{
  BSON binary data subtype constant for md5 data.
}
\seealso{
  \code{\link{mongo.bson.buffer.append.raw}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

