\docType{data}
\name{mongo.binary.old}
\alias{mongo.binary.old}
\title{BSON binary data subtype constant for old format data}
\format{int 2}
\usage{
  mongo.binary.old
}
\value{
  2L
}
\description{
  BSON binary data subtype constant for old format data
  (deprecated).
}
\seealso{
  \code{\link{mongo.bson.buffer.append.raw}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

