% File rmongodb/man/mongo.binary.old.Rd
\name{mongo.binary.old}
\alias{mongo.binary.old}
\title{BSON binary data subtype constant for old format data}
\description{
BSON binary data subtype constant for old format data (deprecated).
}
\usage{
mongo.binary.old
}
\value{
2L
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
