\docType{data}
\name{mongo.binary.user}
\alias{mongo.binary.user}
\title{BSON binary data subtype constant for user data}
\format{\preformatted{ int 128
}}
\usage{
mongo.binary.user
}
\value{
128L
}
\description{
BSON binary data subtype constant for user data.
}
\seealso{
\code{\link{mongo.bson.buffer.append.raw}},\cr
\link{mongo.bson}.
}
\keyword{datasets}

