\docType{data}
\name{mongo.binary.uuid}
\alias{mongo.binary.uuid}
\title{BSON binary data subtype constant for uuid data}
\format{int 3}
\usage{
  mongo.binary.uuid
}
\value{
  4L
}
\description{
  BSON binary data subtype constant for uuid data.
}
\seealso{
  \code{\link{mongo.bson.buffer.append.raw}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

