\docType{data}
\name{mongo.bson.binary}
\alias{mongo.bson.binary}
\title{BSON data type constant for a binary data value}
\format{int 5}
\usage{
  mongo.bson.binary
}
\value{
  5L
}
\description{
  \code{\link{mongo.bson.iterator.type}()} and
  \code{\link{mongo.bson.iterator.next}()} will return this
  constant (5L) to indicate that the value pointer to by an
  iterator is binary data.
}
\seealso{
  \code{\link{mongo.bson.iterator.type}},\cr
  \code{\link{mongo.bson.iterator.next}},\cr
  \link{mongo.bson.iterator},\cr \link{mongo.bson}.
}
\keyword{datasets}

