% File rmongodb/man/mongo.bson.bool.Rd
\name{mongo.bson.bool}
\alias{mongo.bson.bool}
\title{BSON data type constant for a bool value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(8L) to indicate that the value pointer to by an iterator is a bool.
}
\usage{
mongo.bson.bool
}
\value{
8L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson.iterator},\cr
\link{mongo.bson}.
}
