\docType{data}
\name{mongo.bson.code}
\alias{mongo.bson.code}
\title{BSON data type constant for a code value}
\format{int 13}
\usage{
  mongo.bson.code
}
\value{
  13L
}
\description{
  \code{\link{mongo.bson.iterator.type}()} and
  \code{\link{mongo.bson.iterator.next}()} will return this
  constant (13L) to indicate that the value pointer to by
  an iterator is javascript code.
}
\seealso{
  \code{\link{mongo.bson.iterator.type}},\cr
  \code{\link{mongo.bson.iterator.next}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

