% File rmongodb/man/mongo.bson.code.Rd
\name{mongo.bson.code}
\alias{mongo.bson.code}
\title{BSON data type constant for a code value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(13L) to indicate that the value pointer to by an iterator is javascript code.
}
\usage{
mongo.bson.code
}
\value{
13L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
