% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.bson.code.w.scope}
\alias{mongo.bson.code.w.scope}
\title{BSON data type constant for a code with scope value}
\format{\preformatted{ int 15
}}
\usage{
mongo.bson.code.w.scope
}
\value{
15L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this constant (15L) to
indicate that the value pointer to by an iterator is a javascript with a
scope.
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr \link{mongo.bson}
}
\keyword{datasets}

