\docType{data}
\name{mongo.bson.date}
\alias{mongo.bson.date}
\title{BSON data type constant for a date value}
\format{\preformatted{ int 9
}}
\usage{
mongo.bson.date
}
\value{
9L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this
constant (9L) to indicate that the value pointer to by an
iterator is a date/time.
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
\keyword{datasets}

