\docType{data}
\name{mongo.bson.double}
\alias{mongo.bson.double}
\title{BSON data type constant for a double value}
\format{\preformatted{ int 1
}}
\usage{
mongo.bson.double
}
\value{
1L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this
constant (1L) to indicate that the value pointer to by an
iterator is a double.
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
\keyword{datasets}

