% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.bson.eoo}
\alias{mongo.bson.eoo}
\title{BSON data type constant for 'End Of Object'}
\format{\preformatted{ int 0
}}
\usage{
mongo.bson.eoo
}
\value{
0L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this constant (0L) at
the end of the object when there are no more fields through which to
iterate.
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr \link{mongo.bson}.
}
\keyword{datasets}

