% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.bson.int}
\alias{mongo.bson.int}
\title{BSON data type constant for a integer value}
\format{\preformatted{ int 16
}}
\usage{
mongo.bson.int
}
\value{
16L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this constant (16L) to
indicate that the value pointer to by an iterator is a integer (32-bit).
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr \link{mongo.bson}.
}
\keyword{datasets}

