% File rmongodb/man/mongo.bson.long.Rd
\name{mongo.bson.long}
\alias{mongo.bson.long}
\title{BSON data type constant for a long value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(18L) to indicate that the value pointer to by an iterator is a long integer (64 bits).
}
\usage{
mongo.bson.long
}
\value{
18L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
