\docType{data}
\name{mongo.bson.object}
\alias{mongo.bson.object}
\title{BSON data type constant for a subobject value}
\format{int 3}
\usage{
  mongo.bson.object
}
\value{
  3L
}
\description{
  \code{\link{mongo.bson.iterator.type}()} and
  \code{\link{mongo.bson.iterator.next}()} will return this
  constant (3L) to indicate that the value pointer to by an
  iterator is a subobject.
}
\seealso{
  \code{\link{mongo.bson.iterator.type}},\cr
  \code{\link{mongo.bson.iterator.next}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

