\docType{data}
\name{mongo.bson.oid}
\alias{mongo.bson.oid}
\title{BSON data type constant for a oid value}
\format{\preformatted{ int 7
}}
\usage{
mongo.bson.oid
}
\value{
7L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this
constant (7L) to indicate that the value pointer to by an
iterator is a oid (Object ID).
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
\keyword{datasets}

