% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.bson.regex}
\alias{mongo.bson.regex}
\title{BSON data type constant for a regex value}
\format{\preformatted{ int 11
}}
\usage{
mongo.bson.regex
}
\value{
11L
}
\description{
\code{\link{mongo.bson.iterator.type}()} and
\code{\link{mongo.bson.iterator.next}()} will return this constant (11L) to
indicate that the value pointer to by an iterator is a regular expression.
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr \link{mongo.bson}.
}
\keyword{datasets}

