% File rmongodb/man/mongo.bson.symbol.Rd
\name{mongo.bson.symbol}
\alias{mongo.bson.symbol}
\title{BSON data type constant for a symbol value}
\description{
\code{\link{mongo.bson.iterator.type}()} and \code{\link{mongo.bson.iterator.next}()} will return this constant
(14L) to indicate that the value pointer to by an iterator is a symbol.
}
\usage{
mongo.bson.symbol
}
\value{
14L
}
\seealso{
\code{\link{mongo.bson.iterator.type}},\cr
\code{\link{mongo.bson.iterator.next}},\cr
\link{mongo.bson}.
}
