\docType{data}
\name{mongo.bson.timestamp}
\alias{mongo.bson.timestamp}
\title{BSON data type constant for a timestamp value}
\format{int 17}
\usage{
  mongo.bson.timestamp
}
\value{
  17L
}
\description{
  \code{\link{mongo.bson.iterator.type}()} and
  \code{\link{mongo.bson.iterator.next}()} will return this
  constant (17L) to indicate that the value pointer to by
  an iterator is a timestamp.
}
\seealso{
  \code{\link{mongo.bson.iterator.type}},\cr
  \code{\link{mongo.bson.iterator.next}},\cr
  \link{mongo.bson}.
}
\keyword{datasets}

