% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.find.cursor.tailable}
\alias{mongo.find.cursor.tailable}
\title{mongo.find flag constant - cursor tailable}
\format{\preformatted{ int 2
}}
\usage{
mongo.find.cursor.tailable
}
\value{
2L
}
\description{
\code{\link{mongo.find}()} flag constant - cursor tailable.
}
\keyword{datasets}

