% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.find.no.cursor.timeout}
\alias{mongo.find.no.cursor.timeout}
\title{mongo.find flag constant - no cursor timeout}
\format{\preformatted{ int 16
}}
\usage{
mongo.find.no.cursor.timeout
}
\value{
16L
}
\description{
\code{\link{mongo.find}()} flag constant - no cursor timeout.
}
\keyword{datasets}

