% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.index.drop.dups}
\alias{mongo.index.drop.dups}
\title{mongo.index.create flag constant - drop duplicate keys}
\format{\preformatted{ int 4
}}
\usage{
mongo.index.drop.dups
}
\value{
4L
}
\description{
\code{\link{mongo.index.create}()} flag constant - drop duplicate keys.
}
\keyword{datasets}

