\docType{data}
\name{mongo.index.unique}
\alias{mongo.index.unique}
\title{mongo.index.create flag constant - unique keys}
\format{\preformatted{ int 1
}}
\usage{
mongo.index.unique
}
\value{
1L
}
\description{
\code{\link{mongo.index.create}()} flag constant - unique
keys (no duplicates).
}
\keyword{datasets}

