% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mongo.update.basic}
\alias{mongo.update.basic}
\title{mongo.update() flag constant for performing a basic update}
\format{\preformatted{ int 4
}}
\usage{
mongo.update.basic
}
\value{
4L
}
\description{
Flag to \code{\link{mongo.update}()} (4L): Perform a basic update.
}
\seealso{
\code{\link{mongo.update}},\cr \code{\link{mongo.update.multi}}\cr
\code{\link{mongo.update.upsert}}
}
\keyword{datasets}

