\docType{data}
\name{mongo.update.multi}
\alias{mongo.update.multi}
\title{mongo.update() flag constant for updating multiple records}
\format{int 2}
\usage{
  mongo.update.multi
}
\value{
  2L
}
\description{
  Flag to \code{\link{mongo.update}()} (2L): Update
  multiple records rather than just the first one matched
  by criteria.
}
\seealso{
  \code{\link{mongo.update}},\cr
  \code{\link{mongo.update.upsert}},\cr
  \code{\link{mongo.update.basic}}.
}
\keyword{datasets}

