\name{rmongodb-package}
\alias{rmongodb-package}
\alias{rmongodb}
\docType{package}
\title{
R-MongoDB driver
}
\description{
Provides an interface to MongoDB for R
}
\details{
\tabular{ll}{
Package: \tab rmongodb\cr
Type: \tab Package\cr
Version: \tab 1.0.5\cr
Date: \tab 2012-12-22\cr
License: \tab Apache License 2.0\cr
LazyLoad: \tab yes\cr
}
Overview
}
\author{
Gerald Lindsly & 10gen, Inc. (mongo-c-driver)

Maintainer: Gerald Lindsly
gerald.lindsly@gmail.com
}
\seealso{
\link{mongo}
}
\references{
http://www.mongodb.org
}
\keyword{ package }
