\docType{package}
\name{rmongodb-package}
\alias{rmongodb}
\alias{rmongodb-package}
\title{R-MongoDB driver}
\description{
  Provides an interface to MongoDB for R
}
\details{
  \tabular{ll}{ Package: \tab rmongodb\cr Type: \tab
  Package\cr Version: \tab 1.0.5\cr Date: \tab
  2012-12-22\cr License: \tab Apache License 2.0\cr
  LazyLoad: \tab yes\cr } Overview
}
\author{
  Gerald Lindsly & MongoDB, Inc. (mongo-c-driver) & Markus
  Schmidberger (mongosoup.de)

  Maintainer: Markus Schmidberger <markus
}
\references{
  http://www.mongodb.org
}
\seealso{
  \link{mongo}
}
\keyword{package}

