% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{zips}
\alias{Dataset}
\alias{zips}
\title{zips Dataset}
\format{The _id field holds the zip code as a string.

The city field holds the city.

The state field holds the two letter state abbreviation.

The pop field holds the population.

The loc field holds the location as a latitude longitude pair.}
\source{
\url{http://media.mongodb.org/zips.json}
}
\description{
A dataset with US zip data provided by mongodb for education.
}
\examples{
\dontrun{
# code to create the dataset
library(RJSONIO)

json_file <- "http://media.mongodb.org/zips.json"

rL <- readLines(json_file)

zips <- do.call(rbind,lapply(rL,fromJSON))

save(zips, file="data/zips.rda", compress="xz")
}
}

